Aenah - added log out button and beginning functionality for Work Manager. Fixed log in registration more.

+++ app/src/main/java/ututor/edu/csulb/ututor/LogIn.java
@@ -9,6 +9,7 @@ import android.graphics.drawable.Drawable;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.util.Base64;
+import android.util.Log;
 import android.view.View;
 import android.view.animation.AlphaAnimation;
 import android.view.animation.Animation;
@@ -39,6 +40,7 @@ public class LogIn extends AppCompatActivity{
     public ImageView mLogo;
     public TextView mTitle;
     public TextView mForgot;
+    public TextView mAlert;
 
 
     @Override
@@ -53,6 +55,7 @@ public class LogIn extends AppCompatActivity{
         mRemember = (CheckBox) findViewById(R.id.rememberMe);
         mLogo = (ImageView) findViewById(R.id.logo);
         mTitle = (TextView) findViewById(R.id.titleLogo);
+        mAlert = (TextView) findViewById(R.id.alertPassword);
         //mForgot = (TextView) findViewById(R.id.forgotPassword);
 
         Animation expandIn = AnimationUtils.loadAnimation(this, R.anim.expand);
@@ -77,26 +80,20 @@ public class LogIn extends AppCompatActivity{
         boolean checkSaved = sp.getBoolean("isSaved", false);
 
         if(checkSaved == true){
-
-            Toast.makeText(LogIn.this, "Loading Saved Info", Toast.LENGTH_LONG).show();
             String insertEmail = sp.getString("Email", null);
             String insertPassword = sp.getString("Password", null);
 
             mEmail.setText(insertEmail);
             mPassword.setText(insertPassword);
-
             mRemember.setChecked(true);
         }
-
 
 
 
         mSignUp.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-
-                // notify user of success
-                // Intent
+                // Intent to Registration Activity
                 startActivity(new Intent(LogIn.this, Registration.class));
                 finish();
             }
@@ -106,50 +103,64 @@ public class LogIn extends AppCompatActivity{
         mSignIn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
+                // first set alert invisible
+                mAlert.setVisibility(View.INVISIBLE);
 
                 // validate email & password, so access database
+                String databasePassword = "yes";
 
-                // if remember me is selected
-                if(mRemember.isChecked()){
-                    Toast.makeText(LogIn.this, "Saving log-in info", Toast.LENGTH_LONG).show();
-                    String sEmail = mEmail.getText().toString();
-                    String sPass = mPassword.getText().toString();
+                if(mEmail.getText().toString().equals("")){ // if email is blank
+                    mAlert.setVisibility(View.VISIBLE);
+                    Animation shake = AnimationUtils.loadAnimation(LogIn.this, R.anim.shake);
+                    mAlert.startAnimation(shake);
 
-                    SharedPreferences pref = getApplicationContext().getSharedPreferences("TEAM_ANDROID", MODE_PRIVATE);
-                    SharedPreferences.Editor editor = pref.edit();
+                } else if(mPassword.getText().toString().equals(databasePassword)){
 
-                    editor.putBoolean("isSaved", true);
+                    // if remember me is selected
+                    if(mRemember.isChecked()){
+                        String sEmail = mEmail.getText().toString();
+                        String sPass = mPassword.getText().toString();
 
-                    editor.putString("Email", sEmail);
-                    editor.putString("Password", sPass);
+                        SharedPreferences pref = getApplicationContext().getSharedPreferences("TEAM_ANDROID", MODE_PRIVATE);
+                        SharedPreferences.Editor editor = pref.edit();
+                        editor.putBoolean("isSaved", true);
+                        editor.putString("Email", sEmail);
+                        editor.putString("Password", sPass);
+                        editor.commit();
+                    }
+                    else{
 
-                    editor.commit();
+                        SharedPreferences pref = getApplicationContext().getSharedPreferences("TEAM_ANDROID", MODE_PRIVATE);
+                        SharedPreferences.Editor editor = pref.edit();
+                        editor.clear();
+                        editor.apply();
+
+                    }
+
+                    //upload server info about User to the User class
+                    // get info from Database and set them to the User class
+                    // for now, I've just set random values
+                    User cUser = new User();
+                    cUser.setEmail(mEmail.getText().toString());
+                    cUser.setTutor(true);
+                    cUser.setFirstName("[First Name]");
+                    cUser.setLastName("[Last Name]");
+                    Bitmap bitmap = BitmapFactory.decodeResource(getResources(), R.drawable.ututorlogo); // drawable to bitmap
+                    cUser.setProfilePic(bitmap);
+
+                    // send user info to HomePage
+                    Intent i = new Intent(LogIn.this, HomePage.class);
+                    i.putExtra("currentUser", cUser);
+                    startActivity(i);
+                    finish();
+
                 }
-                else{
-
-                    SharedPreferences pref = getApplicationContext().getSharedPreferences("TEAM_ANDROID", MODE_PRIVATE);
-                    SharedPreferences.Editor editor = pref.edit();
-                    editor.clear();
-                    editor.apply();
-
+                else{ // password is wrong, give user notification.
+                    mAlert.setVisibility(View.VISIBLE);
+                    Animation shake = AnimationUtils.loadAnimation(LogIn.this, R.anim.shake);
+                    mAlert.startAnimation(shake);
                 }
 
-
-                //upload server info about User to the User class
-                // get info from Database and set them to the User class
-                // for now, I've just set random values
-                User cUser = new User();
-                cUser.setEmail(mEmail.getText().toString());
-                cUser.setTutor(true);
-                //cUser.setFirstName(String);
-                //cUser.setLastName(String);
-                Bitmap bitmap = BitmapFactory.decodeResource(getResources(), R.drawable.ututorlogo); // drawable to bitmap
-                cUser.setProfilePic(bitmap);
-
-                // send user info to HomePage
-                Intent i = new Intent(LogIn.this, HomePage.class);
-                i.putExtra("currentUser", cUser);
-                startActivity(i);
 
             }
         });
@@ -158,33 +169,32 @@ public class LogIn extends AppCompatActivity{
     }
 
 
+    /**
+     * Temporary Skip button until database is created
+     * @param view
+     */
     public void byPassLogIn(View view){
 
-        //validate username and email
+        //validate username and email in the database
 
 
-        //set random variables
+        // TEMPORARY - set random variables
         mEmail.setText("aenah.ramones@gmail.com");
         mPassword.setText("1234");
 
         // if remember me is selected
         if(mRemember.isChecked()){
-            Toast.makeText(LogIn.this, "Saving log-in info", Toast.LENGTH_LONG).show();
             String sEmail = mEmail.getText().toString();
             String sPass = mPassword.getText().toString();
 
             SharedPreferences pref = getApplicationContext().getSharedPreferences("TEAM_ANDROID", MODE_PRIVATE);
             SharedPreferences.Editor editor = pref.edit();
-
             editor.putBoolean("isSaved", true);
-
             editor.putString("Email", sEmail);
             editor.putString("Password", sPass);
-
             editor.commit();
         }
         else{
-
             SharedPreferences pref = getApplicationContext().getSharedPreferences("TEAM_ANDROID", MODE_PRIVATE);
             SharedPreferences.Editor editor = pref.edit();
             editor.clear();
@@ -195,7 +205,7 @@ public class LogIn extends AppCompatActivity{
         //upload server info about User to the User class
         User cUser = new User();
         cUser.setEmail(mEmail.getText().toString());
-        cUser.setTutor(true);
+        cUser.setTutor(false);
         cUser.setFirstName("Aenah");
         cUser.setLastName("Ramones");
 
@@ -206,6 +216,7 @@ public class LogIn extends AppCompatActivity{
         Intent i = new Intent(LogIn.this, HomePage.class);
         i.putExtra("currentUser", cUser);
         startActivity(i);
+        finish();
 
 
     }
